% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sby_internal_compute_correlation_fortran.R
\name{sby_internal_compute_correlation_fortran}
\alias{sby_internal_compute_correlation_fortran}
\title{Internal Helper: Compute Correlation Fortran}
\usage{
sby_internal_compute_correlation_fortran(numeric_matrix, threshold)
}
\arguments{
\item{numeric_matrix}{A double numeric matrix}

\item{threshold}{A numeric scalar in \verb{[0, 1]}}
}
\value{
A character vector of removed columns
}
\description{
Compute correlation and pruning through native Fortran backend
}
\examples{
sample_matrix <- matrix(stats::rnorm(20), ncol = 2)
colnames(sample_matrix) <- c("A", "B")
sby_internal_compute_correlation_fortran(numeric_matrix = sample_matrix, threshold = 0.9)
}
