% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sby_internal_restore_selected_data.R
\name{sby_internal_restore_selected_data}
\alias{sby_internal_restore_selected_data}
\title{Internal Helper: Restore Selected Data}
\usage{
sby_internal_restore_selected_data(selected_data, original)
}
\arguments{
\item{selected_data}{Filtered tabular object produced by selectors}

\item{original}{Original input object used to infer output class}
}
\value{
A restored tabular object compatible with original class
}
\description{
Restore filtered tabular data to a structure compatible with the
original input class
}
\details{
Matrix inputs are restored as matrices with row names when present.
Tibble inputs are restored as tibbles when the namespace is present.
All other inputs return a base data frame
}
\examples{
original_data <- data.frame(a = 1:3, b = 3:1)
filtered_data <- original_data[, "a", drop = FALSE]
sby_internal_restore_selected_data(filtered_data, original_data)
}
\references{
R CORE TEAM. Writing R Extensions. Vienna: R Foundation.
\href{https://cran.r-project.org/doc/manuals/r-release/R-exts.html}{R Extensions Manual}
}
\seealso{
\code{\link[=sby_select_correlation]{sby_select_correlation()}}, \code{\link[=sby_select_modal_frequency]{sby_select_modal_frequency()}}
}
