% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sby_internal_subset_columns.R
\name{sby_internal_subset_columns}
\alias{sby_internal_subset_columns}
\title{Internal Helper: Subset Tabular Columns}
\usage{
sby_internal_subset_columns(.data, columns)
}
\arguments{
\item{.data}{A data frame, tibble, data.table, or matrix}

\item{columns}{Integer or character column indexes to keep}
}
\value{
A tabular object containing the requested columns
}
\description{
Subset columns consistently across data frames, tibbles, data.tables, and matrices
}
