% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sby_internal_table_row_width.R
\name{sby_internal_table_row_width}
\alias{sby_internal_table_row_width}
\title{Estimate Average Row Width for Parquet Writing}
\usage{
sby_internal_table_row_width(.data)
}
\arguments{
\item{.data}{Object of class \code{data.frame}, \code{tibble}, or \code{data.table}.}
}
\value{
Numeric value with the approximate average row width in bytes.
}
\description{
Estimates the approximate average row width to define the row group size used
during Parquet writing.
}
\details{
The estimate is approximate and is used to choose \code{chunk_size}. For
variable-width columns, such as character columns, the function uses a
deterministic row sample to avoid excessive overhead during writing.
}
\keyword{internal}
