% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sby_table_internal_compression.R
\name{sby_table_internal_compression}
\alias{sby_table_internal_compression}
\title{Define Parquet Compression Codec}
\usage{
sby_table_internal_compression()
}
\value{
Selected codec name.
}
\description{
Automatically defines the compression codec with an emphasis on read and
write speed.
}
\details{
The \code{snappy} codec is preferred when available. Otherwise, the function
uses \code{uncompressed}, which is always supported.

The decision can be overridden with
\code{options(sby_parquet_compression = ...)}.
}
\keyword{internal}
