% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sby_table_internal_normalize_file.R
\name{sby_table_internal_normalize_file}
\alias{sby_table_internal_normalize_file}
\title{Normalize a Parquet File Path}
\usage{
sby_table_internal_normalize_file(file)
}
\arguments{
\item{file}{Complete output file path.}
}
\value{
File path with the \code{.parquet} extension.
}
\description{
Normalizes the output path to ensure that the file extension is
\code{.parquet}.
}
\keyword{internal}
