% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sby_internal_validate_tabular_input.R
\name{sby_internal_validate_tabular_input}
\alias{sby_internal_validate_tabular_input}
\title{Internal Helper: Validate Tabular Input}
\usage{
sby_internal_validate_tabular_input(.data)
}
\arguments{
\item{.data}{Candidate tabular object}
}
\value{
The validated input object
}
\description{
Validate supported tabular input classes for selectors
}
